requirements.txt
asgiref==3.10.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
django-cors-headers==4.9.0
django-filter==23.5
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
idna==3.10
psycopg2-binary==2.9.11
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
